Курсор by_line : построчный ввод из потока
пока только по строкам

Close #1269
