+++ tests/cursor/by_line.cpp
@@ -1,0 +1,37 @@
+/* (c) 2024 Галушин Павел Викторович, galushin@gmail.com
+
+Данный файл -- часть библиотеки SAGA.
+
+SAGA -- это свободной программное обеспечение: вы можете перераспространять ее и/или изменять ее
+на условиях Стандартной общественной лицензии GNU в том виде, в каком она была опубликована Фондом
+свободного программного обеспечения; либо версии 3 лицензии, либо (по вашему выбору) любой более
+поздней версии.
+
+Это программное обеспечение распространяется в надежде, что оно будет полезной, но БЕЗО ВСЯКИХ
+ГАРАНТИЙ; даже без неявной гарантии ТОВАРНОГО ВИДА или ПРИГОДНОСТИ ДЛЯ ОПРЕДЕЛЕННЫХ ЦЕЛЕЙ.
+Подробнее см. в Стандартной общественной лицензии GNU.
+
+Вы должны были получить копию Стандартной общественной лицензии GNU вместе с этим программным
+обеспечение. Если это не так, см. https://www.gnu.org/licenses/.
+*/
+
+// Тестируемый файл
+#include <saga/cursor/by_line.hpp>
+
+// Инфраструктура тестирования
+#include <catch2/catch_amalgamated.hpp>
+
+// Используемые возможности
+#include <saga/cursor/to.hpp>
+
+// Тесты
+TEST_CASE("by_line with delimeter")
+{
+    std::istringstream input("192:168:0:1");
+    std::vector<std::string> const expected{"192", "168", "0", "1"};
+
+    auto const actual = saga::cursor::by_line(input, ':')
+                      | saga::cursor::to<std::vector>();
+
+    CHECK(actual == expected);
+}
